+++ src/main/java/at/sv/hue/api/ManualOverrideTracker.java
@@ -8,11 +8,17 @@ public interface ManualOverrideTracker {
 
     void onLightTurnedOn(String id);
 
+    void onLightTurnedOnBySyncedScene(String id);
+
+    void onLightTurnedOnManually(String id);
+
     void onLightOff(String id);
 
     boolean isOff(String id);
 
     boolean wasJustTurnedOn(String id);
+
+    boolean wasTurnedOnBySyncedScene(String id);
 
     void onAutomaticallyAssigned(String id);
 }